[AUN] aun11-1 handler
