+++ lib/aunHandler.js
@@ -30,6 +30,7 @@ var aun2_1Control     = require('../lib/qa/aun2_1Handler');
 var aun3_3Control     = require('../lib/qa/aun3_3Handler');
 var aun6_1Control     = require('../lib/qa/aun6_1Handler');
 var aun6_2Control     = require('../lib/qa/aun6_2Handler');
+var aun11_1Control    = require('../lib/qa/aun11_1Handler');
 
 
 var isLoggedIn = require('middleware/loginChecker');
@@ -41,6 +42,7 @@ app.use('/aun2-1', aun2_1Control  ); //aun2_1
 app.use('/aun3-3', aun3_3Control  ); //aun3_3
 app.use('/aun6-1', aun6_1Control  ); //aun6_1
 app.use('/aun6-2', aun6_2Control  ); //aun6_2
+app.use('/aun11-1', aun11_1Control  ); //aun6_2
