+++ views/headerqa.ejs
@@ -232,7 +232,7 @@ $('#cssmenu li.has-sub>a').on('click', function(){
 
                 <li class='has-sub'><a href='#'><span>AUN 11: Quality Assurence of Teching and Learning Process</span></a>
                     <ul>
-                        <li><a href="/aun11-1?program={{programname}}&year={{year}}&acid={{acid}}">AUN 11.1: The curriculum is developed by all teaching staff members</a></li>
+                        <li><a href="/aun/aun11-1?program=<%=programname%>&year=<%=year%>&acid=<%=acid%>">AUN 11.1: The curriculum is developed by all teaching staff members</a></li>
                         <li><a href="/aun11-4?program={{programname}}&year={{year}}&acid={{acid}}">AUN 11.4: The curriculum is regurly evaluated at reasonable time periods</a></li>
                         
                     </ul>
