+++ views/qa/editqa/aun11.1_adddevcom.hbs
@@ -53,7 +53,7 @@
 </script>
 
 
-<form class="form-horizontal" role="form" action="/add_aun11-1?program={{program}}&year={{year}}" method="post">
+<form class="form-horizontal" role="form" action="/aun/aun11-1/add_aun11-1?program={{programname}}&year={{year}}&acid={{acid}}" method="post">
     <h3>Edit Program development committe</h3>
     <p hidden>
