+++ views/qa/qa-aun11.1.hbs
@@ -58,7 +58,7 @@
                         {{/each}}
                     </tbody>
 					{{/each}}
-                    <a href = "/add_aun11-1?program={{program}}&year={{year}}"><button id="tab_2_submit" class="btn btn-mini" type="button">Add</button>
+                    <a href = "/aun/aun11-1/add_aun11-1?program={{programname}}&year={{year}}&acid={{acid}}"><button id="tab_2_submit" class="btn btn-mini" type="button">Add</button>
                   </table>
 
                 </div>
